Archive/server/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
flask_cors==4.0.0
python-dotenv==1.0.1
requests==2.26.0